+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.5.6
 
+* `Ratings`:
+    * `GC`:
+        * Now GC will autoclear ratings even if post is absent in posts repo
+
 ## 0.5.5
 
 * Dependencies update